#29: Provide an AMI for arm64
* initial version
* fallback to 'x64' when 2nd positional parameter (ARCH) is unset or null
* shift parameter used to allow root login via SSH to 3rd position
* consolidated parameters 'XLT_VERSION' and 'SOURCE' to 'XLT_SOURCE'
* allow user to configure size of root volume when building Amazon images

+++ scripts/common/xlt-home/setup-image.sh
@@ -1,10 +1,30 @@
 #!/bin/bash
 set -e
 
+## Script parameters ##
+
+XLT_SOURCE="$1"
+ARCH="${2:-x64}"
+ENABLE_ROOT_LOGIN="$3"
+
+# Check if script is run as root user
 if ! test `id -u` -eq 0 ; then
   echo "This script must be run as root"
   exit 1
 fi
+
+if [ -z "$XLT_SOURCE" ]; then
+  echo "Setup script must be invoked with XLT-Version or path to XLT distribution archive as 1st argument"
+  exit 1
+fi
+
+if [ "$ARCH" != "x64" -a "$ARCH" != "arm64" ]; then
+  echo "Unsupported architecture: \"$ARCH\""
+  exit 1
+fi
+
+echo "Building image for architecture \"$ARCH\" ..."
+
 
 ## first the variables
 SCRIPT_DIR=`realpath -m $0/..`
@@ -13,6 +33,7 @@ INIT_SCRIPT_DIR="$(dirname $SCRIPT_DIR)/init.d"
 XLT_USER="xlt"
 XLT_HOME="/home/$XLT_USER"
 XLT_WORKDIR="/mnt/$XLT_USER"
+TARGET_ARCHIVE="$XLT_HOME/xlt.zip"
 
 JAVA_HOME=/usr/lib/jvm/java-11-openjdk
 
@@ -24,14 +45,20 @@ XLT_INITD_SCRIPT_NAME="xlt"
 XLT_START_SCRIPT_NAME="start-xlt.sh"
 NTP_START_SCRIPT="ntptime"
 
-FIREFOX_ESR_VERSION="102.0esr"
-FIREFOX_ESR_DOWNLOAD_URL="https://download-installer.cdn.mozilla.net/pub/firefox/releases/${FIREFOX_ESR_VERSION}/linux-x86_64/en-US/firefox-${FIREFOX_ESR_VERSION}.tar.bz2"
-FIREFOX_ESR_CHECKSUM="225b5170d80ebedb9c0477a45026f617f4e2bb4d2cd3cdfa1822f8e0c6adff49"
-GECKODRIVER_VERSION="v0.31.0"
-GECKODRIVER_DOWNLOAD_URL="https://github.com/mozilla/geckodriver/releases/download/${GECKODRIVER_VERSION}/geckodriver-${GECKODRIVER_VERSION}-linux64.tar.gz"
+if [ "$ARCH" == "arm64" ]; then
+  # Mozilla does not provide prebuilt packages of Firefox-ESR/geckodriver for ARM -> have to install them via package manager
+  OPENJDK_DOWNLOAD_URL="https://github.com/adoptium/temurin11-binaries/releases/download/jdk-11.0.16.1%2B1/OpenJDK11U-jdk_aarch64_linux_hotspot_11.0.16.1_1.tar.gz"
+  OPENJDK_CHECKSUM="2b89cabf0ce1c2cedadd92b798d6e9056bc27c71a06f5ba24ede5dc9c316e3e8"
+else
+  FIREFOX_ESR_VERSION="102.0esr"
+  FIREFOX_ESR_DOWNLOAD_URL="https://download-installer.cdn.mozilla.net/pub/firefox/releases/${FIREFOX_ESR_VERSION}/linux-x86_64/en-US/firefox-${FIREFOX_ESR_VERSION}.tar.bz2"
+  FIREFOX_ESR_CHECKSUM="225b5170d80ebedb9c0477a45026f617f4e2bb4d2cd3cdfa1822f8e0c6adff49"
+  GECKODRIVER_VERSION="v0.31.0"
+  GECKODRIVER_DOWNLOAD_URL="https://github.com/mozilla/geckodriver/releases/download/${GECKODRIVER_VERSION}/geckodriver-${GECKODRIVER_VERSION}-linux64.tar.gz"
 
-OPENJDK_DOWNLOAD_URL="https://github.com/adoptium/temurin11-binaries/releases/download/jdk-11.0.15%2B10/OpenJDK11U-jdk_x64_linux_hotspot_11.0.15_10.tar.gz"
-OPENJDK_CHECKSUM="5fdb4d5a1662f0cca73fec30f99e67662350b1fa61460fa72e91eb9f66b54d0b"
+  OPENJDK_DOWNLOAD_URL="https://github.com/adoptium/temurin11-binaries/releases/download/jdk-11.0.15%2B10/OpenJDK11U-jdk_x64_linux_hotspot_11.0.15_10.tar.gz"
+  OPENJDK_CHECKSUM="5fdb4d5a1662f0cca73fec30f99e67662350b1fa61460fa72e91eb9f66b54d0b"
+fi
 
 ## check referenced files existance
 function checkFile {
@@ -128,40 +155,47 @@ chmod +x /etc/profile.d/jdk.sh
 sed -i -e 's/^\(keystore\.type\)=.*$/\1=JKS/' $JAVA_HOME/conf/security/java.security
 
 # Install Root CA certs for Java
-
-apt-get install -y --no-install-recommends ca-certificates-java \
+DEBIAN_FRONTEND=noninteractive apt-get --no-install-recommends install -y ca-certificates-java \
   && rm $JAVA_HOME/lib/security/cacerts \
   && ln -s /etc/ssl/certs/java/cacerts $JAVA_HOME/lib/security/
-
 
 # Install Maven (Maven needs Java, so install it in the correct order)
 DEBIAN_FRONTEND=noninteractive apt-get --no-install-recommends -y install maven
 
-
-# Download Firefox ESR and put it into path
-curl -L $FIREFOX_ESR_DOWNLOAD_URL -o /tmp/firefox.tar.bz2
-echo "$FIREFOX_ESR_CHECKSUM /tmp/firefox.tar.bz2" | sha256sum -c --status - || exit 1
-tar -xj -C /tmp -f /tmp/firefox.tar.bz2
-[ -d /usr/lib/firefox-esr ] && rm -rf /usr/lib/firefox-esr
-mv /tmp/firefox /usr/lib/firefox-esr
-ln -s /usr/lib/firefox-esr/firefox /usr/bin/firefox-esr
-rm /tmp/firefox.tar.bz2
-
-# Download Geckodriver from GitHub and put it into path
-echo "Install geckodriver"
-curl -L $GECKODRIVER_DOWNLOAD_URL -o /tmp/geckodriver-linux64.tgz
-tar -xz -C /usr/bin -f /tmp/geckodriver-linux64.tgz
-chown root:root /usr/bin/geckodriver
-chmod 755 /usr/bin/geckodriver
-rm /tmp/geckodriver-linux64.tgz
-
-# Download chromedriver from Google and put it into path
-echo "Install chromedriver"
-curl -L $(_chromedriverUrl) -o /tmp/chromedriver_linux64.zip
-unzip -d /usr/bin /tmp/chromedriver_linux64.zip
-chown root:root /usr/bin/chromedriver
-chmod 755 /usr/bin/chromedriver
-rm /tmp/chromedriver_linux64.zip
+# Install Firefox-ESR, geckodriver, and chromedriver
+if [ "$ARCH" == "arm64" ]; then
+  # install via package manager as no download available
+  # TODO: firefox-esr?
+  echo "Install geckodriver + chromedriver via APT"
+  DEBIAN_FRONTEND=noninteractive apt-get --no-install-recommends -y install \
+    chromium-chromedriver \
+    firefox-geckodriver
+else
+  # Download Firefox ESR and put it into path
+  curl -L $FIREFOX_ESR_DOWNLOAD_URL -o /tmp/firefox.tar.bz2
+  echo "$FIREFOX_ESR_CHECKSUM /tmp/firefox.tar.bz2" | sha256sum -c --status - || exit 1
+  tar -xj -C /tmp -f /tmp/firefox.tar.bz2
+  [ -d /usr/lib/firefox-esr ] && rm -rf /usr/lib/firefox-esr
+  mv /tmp/firefox /usr/lib/firefox-esr
+  ln -s /usr/lib/firefox-esr/firefox /usr/bin/firefox-esr
+  rm /tmp/firefox.tar.bz2
+  
+  # Download Geckodriver from GitHub and put it into path
+  echo "Install geckodriver"
+  curl -L $GECKODRIVER_DOWNLOAD_URL -o /tmp/geckodriver-linux64.tgz
+  tar -xz -C /usr/bin -f /tmp/geckodriver-linux64.tgz
+  chown root:root /usr/bin/geckodriver
+  chmod 755 /usr/bin/geckodriver
+  rm /tmp/geckodriver-linux64.tgz
+  
+  # Download chromedriver from Google and put it into path
+  echo "Install chromedriver"
+  curl -L $(_chromedriverUrl) -o /tmp/chromedriver_linux64.zip
+  unzip -d /usr/bin /tmp/chromedriver_linux64.zip
+  chown root:root /usr/bin/chromedriver
+  chmod 755 /usr/bin/chromedriver
+  rm /tmp/chromedriver_linux64.zip
+fi
 
 # Setup XLT start script
 echo "Install XLT start script"
@@ -239,7 +273,7 @@ fi
 
 ## secure login
 echo "Secure login"
-if [ "$2" = "enableRootLogin" ]
+if [ "$ENABLE_ROOT_LOGIN" = "enableRootLogin" ]
 then
   echo "Enable root login"
   sed -ri 's/^\s*PermitRootLogin\s*no$/PermitRootLogin\ yes/g' /etc/ssh/sshd_config
@@ -249,36 +283,34 @@ else
 fi
 
 # remove 'unattended-upgrades'
-apt-get -q -y purge unattended-upgrades
+DEBIAN_FRONTEND=noninteractive apt-get -q -y purge unattended-upgrades
 
 # install XLT
 echo "Installing XLT"
-SOURCE=$1
-TARGET_ARCHIVE="$XLT_HOME/xlt.zip"
 
-if [ ! -f $SOURCE ]; then
-  SOURCE="https://oss.sonatype.org/service/local/artifact/maven/redirect?r=releases&g=com.xceptance&a=xlt&e=zip&v=$SOURCE"
+if [ ! -f $XLT_SOURCE ]; then
+  XLT_SOURCE="https://oss.sonatype.org/service/local/artifact/maven/redirect?r=releases&g=com.xceptance&a=xlt&e=zip&v=$XLT_SOURCE"
 fi
 
-if [[ $SOURCE == http://* ]] || [[ $SOURCE == https://* ]]; then
+if [[ $XLT_SOURCE == http://* ]] || [[ $XLT_SOURCE == https://* ]]; then
   # load from URL
   echo "download ..."
-  curl -s -f -o $TARGET_ARCHIVE -L "$SOURCE"
+  curl -s -f -o "$TARGET_ARCHIVE" -L "$XLT_SOURCE"
 else
   # is not a URL -> must be a file
-  if [ -r "$SOURCE" ] && [ -f "$SOURCE" ]; then
+  if [ -r "$XLT_SOURCE" ] && [ -f "$XLT_SOURCE" ]; then
     # get from file
-    mv $SOURCE $TARGET_ARCHIVE
+    mv "$XLT_SOURCE" "$TARGET_ARCHIVE"
   fi
 fi
 
 if [ ! -f "$TARGET_ARCHIVE" ]; then
-  echo "Given parameter '$SOURCE' is neither a valid URL nor points to an existing file."
+  echo "Given parameter '$XLT_SOURCE' is neither a valid URL nor points to an existing file."
   exit 2;
 fi
 
-echo "set up rights"
-chown xlt:xlt $TARGET_ARCHIVE
+echo "Set up rights"
+chown xlt:xlt "$TARGET_ARCHIVE"
 
 
 # prepare AMI creation
@@ -309,8 +341,8 @@ fi
 ## clean up
 echo "Clean up setup files"
 
-apt-get -y clean && apt-get -y autoremove && rm -rf /var/lib/lists/*
+DEBIAN_FRONTEND=noninteractive apt-get -y clean && apt-get -y autoremove && rm -rf /var/lib/lists/*
 cd /
-rm -rf $SCRIPT_DIR $INIT_SCRIPT_DIR
+rm -rf "$SCRIPT_DIR" "$INIT_SCRIPT_DIR"
 
 echo "Setup finished."
